api: refactoring
